+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-## [1.0.0] 
+## [1.0.2]
+
+* 新增onClickOut()和onClickBack()
+* Added onClickOut () and onClickBack ()
+
+## [1.0.1]
+
+*修改部分方法
+* Modification method.
+
+## [1.0.0]
 
 * first version.
 
-## [1.0.1] 
 
-* Modification method.